requirements.txt
flask
python-dotenv
google-generativeai
sympy
numpy
matplotlib
pillow
opencv-python
requests
